@@ -13,10 +13,12 @@ class Server extends Base
 {
     // Default options
     protected static $default_options = [
-      'timeout'       => null,
+      'filter'        => ['text', 'binary'],
       'fragment_size' => 4096,
-      'port'          => 8000,
       'logger'        => null,
+      'port'          => 8000,
+      'return_obj'    => false,
+      'timeout'       => null,
     ];
 
     protected $addr;
@@ -32,20 +34,28 @@ class Server extends Base
      *   - fragment_size: Set framgemnt size.  Default: 4096
      *   - port:          Chose port for listening. Default 8000.
      */
-    public function __construct(array $options = array())
+    public function __construct(array $options = [])
     {
         $this->options = array_merge(self::$default_options, $options);
         $this->port = $this->options['port'];
         $this->setLogger($this->options['logger']);
 
+        $error = $errno = $errstr = null;
+        set_error_handler(function (int $severity, string $message, string $file, int $line) use (&$error) {
+            $this->logger->warning($message, ['severity' => $severity]);
+            $error = $message;
+        }, E_ALL);
+
         do {
-            $this->listening = @stream_socket_server("tcp://0.0.0.0:$this->port", $errno, $errstr);
+            $this->listening = stream_socket_server("tcp://0.0.0.0:$this->port", $errno, $errstr);
         } while ($this->listening === false && $this->port++ < 10000);
 
+        restore_error_handler();
+
         if (!$this->listening) {
-            $error = "Could not open listening socket: {$errstr} ({$errno})";
+            $error = "Could not open listening socket: {$errstr} ({$errno}) {$error}";
             $this->logger->error($error);
-            throw new ConnectionException($error, $errno);
+            throw new ConnectionException($error, (int)$errno);
         }
 
         $this->logger->info("Server listening to port {$this->port}");
@@ -59,22 +69,22 @@ class Server extends Base
         $this->socket = null;
     }
 
-    public function getPort()
+    public function getPort(): int
     {
         return $this->port;
     }
 
-    public function getPath()
+    public function getPath(): string
     {
         return $this->request_path;
     }
 
-    public function getRequest()
+    public function getRequest(): array
     {
         return $this->request;
     }
 
-    public function getHeader($header)
+    public function getHeader($header): ?string
     {
         foreach ($this->request as $row) {
             if (stripos($row, $header) !== false) {
@@ -85,36 +95,44 @@ class Server extends Base
         return null;
     }
 
-    public function accept()
+    public function accept(): bool
     {
         $this->socket = null;
         return (bool)$this->listening;
     }
 
-    protected function connect()
+    protected function connect(): void
     {
-        if (empty($this->options['timeout'])) {
-            $this->socket = @stream_socket_accept($this->listening);
-            if (!$this->socket) {
-                $error = 'Server failed to connect.';
-                $this->logger->error($error);
-                throw new ConnectionException($error);
-            }
+
+        $error = null;
+        set_error_handler(function (int $severity, string $message, string $file, int $line) use (&$error) {
+            $this->logger->warning($message, ['severity' => $severity]);
+            $error = $message;
+        }, E_ALL);
+
+        if (isset($this->options['timeout'])) {
+            $this->socket = stream_socket_accept($this->listening, $this->options['timeout']);
         } else {
-            $this->socket = @stream_socket_accept($this->listening, $this->options['timeout']);
-            if (!$this->socket) {
-                $error = 'Server failed to connect.';
-                $this->logger->error($error);
-                throw new ConnectionException($error);
-            }
+            $this->socket = stream_socket_accept($this->listening);
+        }
+
+        restore_error_handler();
+
+        if (!$this->socket) {
+            $this->throwException("Server failed to connect. {$error}");
+        }
+        if (isset($this->options['timeout'])) {
             stream_set_timeout($this->socket, $this->options['timeout']);
         }
 
+        $this->logger->info("Client has connected to port {port}", [
+            'port' => $this->port,
+            'pier' => stream_socket_get_name($this->socket, true),
+        ]);
         $this->performHandshake();
-        $this->logger->info("Server connected to port {$this->port}");
     }
 
-    protected function performHandshake()
+    protected function performHandshake(): void
     {
         $request = '';
         do {
@@ -153,5 +171,6 @@ class Server extends Base
                 . "\r\n";
 
         $this->write($header);
+        $this->logger->debug("Handshake on {$get_uri}");
     }
 }
