@@ -20,9 +20,13 @@ class MockSocket
         if ($function == 'get_resource_type' && is_null($current)) {
             return null; // Catch destructors
         }
-        self::$asserter->assertEquals($function, $current['function']);
+        self::$asserter->assertEquals($current['function'], $function);
         foreach ($current['params'] as $index => $param) {
             self::$asserter->assertEquals($param, $params[$index], json_encode([$current, $params]));
+        }
+        if (isset($current['error'])) {
+            $map = array_merge(['msg' => 'Error', 'type' => E_USER_NOTICE], (array)$current['error']);
+            trigger_error($map['msg'], $map['type']);
         }
         if (isset($current['return-op'])) {
             return self::op($current['return-op'], $params, $current['return']);
@@ -34,13 +38,13 @@ class MockSocket
     }
 
     // Check if all expected calls are performed
-    public static function isEmpty()
+    public static function isEmpty(): bool
     {
         return empty(self::$queue);
     }
 
     // Initialize call queue
-    public static function initialize($op_file, $asserter)
+    public static function initialize($op_file, $asserter): void
     {
         $file = dirname(__DIR__) . "/scripts/{$op_file}.json";
         self::$queue = json_decode(file_get_contents($file), true);
